Adiciona LoginForm component para reduzir duplicação de código nos components de Cadastro e Login.
