+++ src/pages/login/Login.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useHistory } from 'react-router';
 import PropTypes from 'prop-types';
 
@@ -7,62 +7,29 @@ import {
   CssBaseline,
   Container,
   Typography,
-  TextField,
   Grid,
   Link,
 } from '@material-ui/core';
 
-import useErros from '../../hooks/useErros';
 import AutenticacaoService from '../../services/AutenticacaoService';
+import CustomLoginForm from '../../components/login-form/LoginForm';
 
-import {
-  LoginContainer,
-  LoginForm,
-  IconAvatar,
-  LoginButton,
-} from '../../assets/GlobalStyles';
+import { LoginContainer, IconAvatar } from '../../assets/GlobalStyles';
 
 const Login = ({ handleChangeLogin }) => {
-  const [usuario, setUsuario] = useState('');
-  const [senha, setSenha] = useState('');
-
   const history = useHistory();
 
-  const validacoes = {
-    usuario: (usuarioAValidar) => {
-      if (!usuarioAValidar.length)
-        return { valido: false, texto: 'Campo obrigatório' };
-      if (usuarioAValidar.length <= 3)
-        return {
-          valido: false,
-          texto: 'Usuário deve ter ao menos 4 caracteres',
-        };
-      return { valido: true };
-    },
-
-    senha: (senhaAValidar) => {
-      if (!senhaAValidar || senhaAValidar.length < 6)
-        return { valido: false, texto: 'Senha deve ter ao menos 6 caracteres' };
-      return { valido: true };
-    },
-  };
-
-  const [erros, validarCampos, possoEnviar] = useErros(validacoes);
-
-  const logar = (e) => {
-    if (possoEnviar) {
-      e.preventDefault();
-      AutenticacaoService.autenticar({ nome: usuario, senha }).then(
-        (response) => {
-          if (!response?.error) {
-            handleChangeLogin({
-              nome: usuario,
-              token: response.token,
-            });
-            history.push('/');
-          }
+  const logar = (usuario, possoEnviar) => {
+    if (possoEnviar()) {
+      AutenticacaoService.autenticar(usuario).then((response) => {
+        if (!response?.error) {
+          handleChangeLogin({
+            nome: usuario.nome,
+            token: response.token,
+          });
+          history.push('/');
         }
-      );
+      });
     }
   };
 
@@ -76,56 +43,7 @@ const Login = ({ handleChangeLogin }) => {
         <Typography component='h1' variant='h5'>
           Login
         </Typography>
-        <LoginForm onSubmit={(e) => logar(e)}>
-          <TextField
-            value={usuario}
-            onBlur={validarCampos}
-            onChange={(e) => {
-              setUsuario(e.target.value);
-              validarCampos(e);
-            }}
-            error={!erros.usuario.valido}
-            helperText={erros.usuario.texto}
-            variant='outlined'
-            margin='normal'
-            id='usuario'
-            name='usuario'
-            label='Usuário'
-            type='text'
-            inputProps={{ 'data-testid': 'usuario' }}
-            fullWidth
-            required
-          />
-          <TextField
-            value={senha}
-            onBlur={validarCampos}
-            onChange={(e) => {
-              setSenha(e.target.value);
-              validarCampos(e);
-            }}
-            error={!erros.senha.valido}
-            helperText={erros.senha.texto}
-            variant='outlined'
-            margin='normal'
-            id='senha'
-            name='senha'
-            label='Senha'
-            type='password'
-            inputProps={{ 'data-testid': 'senha' }}
-            fullWidth
-            required
-          />
-
-          <LoginButton
-            disabled={!possoEnviar()}
-            type='submit'
-            variant='contained'
-            color='primary'
-            fullWidth
-          >
-            Entrar
-          </LoginButton>
-        </LoginForm>
+        <CustomLoginForm modo={'login'} handleChangeForm={logar} />
         <Grid container>
           <Link href='/cadastro'>Não possui conta? Cadastrar</Link>
         </Grid>
